+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on
 [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project
 adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.4] - 2026-04-16
+
+### Changed
+
+- Added dry-run mode to GitOps PR CLI tool.
+- Expanded usage examples in GitOps PR CLI tool.
+- Simplified the end decision logic in the GitOps PR CLI tool to optionally
+  switch back to the main branch.
+
 ## [0.3.3] - 2026-04-16
 
 ### Changed